+++ pkg/artifactory/resource/repository/repository.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/go-resty/resty/v2"
 	"github.com/hashicorp/go-cty/cty"
-	"github.com/hashicorp/terraform-plugin-framework-validators/setvalidator"
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
@@ -511,11 +510,8 @@ var BaseAttributes = map[string]schema.Attribute{
 		Optional:    true,
 		Computed:    true,
 		Default:     setdefault.StaticValue(types.SetValueMust(types.StringType, []attr.Value{})),
-		Validators: []validator.Set{
-			setvalidator.SizeBetween(0, 2),
-		},
-		MarkdownDescription: "Project environment for assigning this repository to. Allow values: \"DEV\", \"PROD\", or one of custom environment. " +
-			"Before Artifactory 7.53.1, up to 2 values (\"DEV\" and \"PROD\") are allowed. From 7.53.1 onward, only one value is allowed. " +
+		MarkdownDescription: "Before Artifactory 7.53.1, up to 2 values (`DEV` and `PROD`) are allowed. From 7.53.1 to 7.107.1, " +
+			"only one value is allowed. From 7.107.1, multiple values are allowed." +
 			"The attribute should only be used if the repository is already assigned to the existing project. If not, " +
 			"the attribute will be ignored by Artifactory, but will remain in the Terraform state, which will create " +
 			"state drift during the update.",
@@ -612,13 +608,11 @@ var BaseSchemaV1 = map[string]*sdkv2_schema.Schema{
 	"project_environments": {
 		Type:     sdkv2_schema.TypeSet,
 		Elem:     &sdkv2_schema.Schema{Type: sdkv2_schema.TypeString},
-		MinItems: 0,
-		MaxItems: 2,
 		Set:      sdkv2_schema.HashString,
 		Optional: true,
 		Computed: true,
-		Description: "Project environment for assigning this repository to. Allow values: \"DEV\", \"PROD\", or one of custom environment. " +
-			"Before Artifactory 7.53.1, up to 2 values (\"DEV\" and \"PROD\") are allowed. From 7.53.1 onward, only one value is allowed. " +
+		Description: "Before Artifactory 7.53.1, up to 2 values (`DEV` and `PROD`) are allowed. From 7.53.1 to 7.107.1, " +
+			"only one value is allowed. From 7.107.1, multiple values are allowed." +
 			"The attribute should only be used if the repository is already assigned to the existing project. If not, " +
 			"the attribute will be ignored by Artifactory, but will remain in the Terraform state, which will create " +
 			"state drift during the update.",