+++ pkg/artifactory/resource/repository/repository_test.go
@@ -329,7 +329,7 @@ func TestAccRepository_invalid_project_environments_after_7_53_1_before_7_107_1(
 	})
 }
 
-func TestAccRepository_can_set_two_project_environments_after_7_107_1(t *testing.T) {
+func TestAccRepository_can_set_multi_project_environments_after_7_107_1(t *testing.T) {
 	projectKey := fmt.Sprintf("t%d", testutil.RandomInt())
 	repoName := fmt.Sprintf("%s-generic-local", projectKey)
 
@@ -340,6 +340,9 @@ func TestAccRepository_can_set_two_project_environments_after_7_107_1(t *testing
 		"projectKey": projectKey,
 	}
 	localRepositoryBasic := util.ExecuteTemplate("TestAccLocalGenericRepository", `
+		resource "artifactory_global_environment" "staging" {
+  			name = "STAGING"
+		}
 		resource "project" "{{ .projectKey }}" {
 			key = "{{ .projectKey }}"
 			display_name = "{{ .projectKey }}"
@@ -357,7 +360,8 @@ func TestAccRepository_can_set_two_project_environments_after_7_107_1(t *testing
 		resource "artifactory_local_generic_repository" "{{ .name }}" {
 		  key                  = "{{ .name }}"
 	 	  project_key          = project.{{ .projectKey }}.key
-	 	  project_environments = ["DEV", "PROD"]
+	 	  project_environments = ["DEV", "PROD", "STAGING"]
+		  depends_on           = [artifactory_global_environment.staging, project.{{ .projectKey }}]
 		}
 	`, params)
 
@@ -379,9 +383,46 @@ func TestAccRepository_can_set_two_project_environments_after_7_107_1(t *testing
 				},
 				Config: localRepositoryBasic,
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestCheckResourceAttr(fqrn, "project_environments.#", "2"),
+					resource.TestCheckResourceAttr(fqrn, "project_environments.#", "3"),
 					resource.TestCheckResourceAttr(fqrn, "project_environments.0", "DEV"),
 					resource.TestCheckResourceAttr(fqrn, "project_environments.1", "PROD"),
+					resource.TestCheckResourceAttr(fqrn, "project_environments.2", "STAGING"),
+				),
+			},
+		},
+	})
+}
+
+func TestAccRepository_can_set_empty_project_environments_after_7_107_1(t *testing.T) {
+	repoName := fmt.Sprintf("test-generic-local-%d", testutil.RandomInt())
+	_, fqrn, name := testutil.MkNames(repoName, "artifactory_local_generic_repository")
+
+	params := map[string]interface{}{
+		"name": name,
+	}
+
+	localRepositoryWithEmptyEnvironments := util.ExecuteTemplate("TestAccLocalGenericRepository", `
+		resource "artifactory_local_generic_repository" "{{ .name }}" {
+		  key                  = "{{ .name }}"
+	 	  project_environments = []
+		}
+	`, params)
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.PreCheck(t) },
+		ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
+		CheckDestroy:             acctest.VerifyDeleted(t, fqrn, "key", acctest.CheckRepo),
+		Steps: []resource.TestStep{
+			{
+				SkipFunc: func() (bool, error) {
+					meta := acctest.Provider.Meta().(util.ProviderMetadata)
+					multiSupport, err := util.CheckVersion(meta.ArtifactoryVersion, repository.MultipleEnvironmentsSupportedVersion)
+					return !multiSupport, err
+				},
+				Config: localRepositoryWithEmptyEnvironments,
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(fqrn, "key", name),
+					resource.TestCheckResourceAttr(fqrn, "project_environments.#", "0"),
 				),
 			},
 		},